@@ -11,9 +11,12 @@
  * granted to it by virtue of its status as an Intergovernmental Organization
  * or submit itself to any jurisdiction.
  */
-const RunDetectorQuality = {
-    Good: 'good',
-    Bad: 'bad',
+
+export const RunQualities = {
+    GOOD: 'good',
+    BAD: 'bad',
+    TEST: 'test',
+    NONE: 'none',
 };
 
-module.exports = RunDetectorQuality;
+export const RUN_QUALITIES = Object.values(RunQualities);
