@@ -71,21 +71,42 @@ class SequelizeDatabase {
     }
 
     /**
-     * Performs connection to the database.
+     * Performs connection to the database, and retry until success
      *
-     * @returns {Promise} Promise object represents the outcome.
+     * @returns {Promise} resolves once the connection has been done
      */
     async connect() {
-        this.logger.info('Starting...');
+        this.logger.debug('Starting bookkeeping database connection');
+        const retryThrottle = 5000;
 
-        try {
-            await this.sequelize.authenticate();
-        } catch (error) {
-            this.logger.error(`Error while starting: ${error}`);
-            return Promise.reject(error);
+        let success = false;
+        let failedOnce = false;
+        while (!success) {
+            const attemptTime = Date.now();
+
+            const authenticationPromise = this.sequelize.authenticate();
+            authenticationPromise.catch((error) => {
+                const retryIn = attemptTime + retryThrottle - Date.now();
+                if (!failedOnce) {
+                    this.logger.error(`Error while starting bookkeeping database connection: ${error}`);
+                    failedOnce = true;
+                }
+
+                if (retryIn > 0) {
+                    this.logger.debug(`New bookkeeping database connection attempt in ${retryIn} ms`);
+                }
+            });
+
+            await Promise.allSettled([
+                authenticationPromise,
+                new Promise((resolve) => setTimeout(() => resolve(), retryThrottle)),
+            ]).then(([{ status }]) => {
+                if (status !== 'rejected') {
+                    success = true;
+                }
+            });
         }
-
-        this.logger.info('Started');
+        this.logger.debug('Bookkeeping database connected');
     }
 
     /**
