@@ -12,6 +12,7 @@
  */
 
 const Sequelize = require('sequelize');
+const { RUN_QUALITIES, RunQualities } = require('../../domain/enums/RunQualities.js');
 
 module.exports = (sequelize) => {
     const Run = sequelize.define('Run', {
@@ -71,7 +72,8 @@ module.exports = (sequelize) => {
             type: Sequelize.STRING,
         },
         runQuality: {
-            type: Sequelize.ENUM('good', 'bad', 'unknown', 'test'),
+            type: Sequelize.ENUM(...RUN_QUALITIES),
+            default: RunQualities.NONE,
         },
         nDetectors: {
             type: Sequelize.INTEGER,
