@@ -11,6 +11,8 @@
  * or submit itself to any jurisdiction.
  */
 
+const { RUN_DETECTOR_QUALITIES, RunDetectorQualities } = require('../../domain/enums/RunDetectorQualities.js');
+
 module.exports = (sequelize) => {
     const Sequelize = require('sequelize');
 
@@ -22,8 +24,8 @@ module.exports = (sequelize) => {
             type: Sequelize.INTEGER,
         },
         quality: {
-            type: Sequelize.ENUM('good', 'bad'),
-            default: 'good',
+            type: Sequelize.ENUM(...RUN_DETECTOR_QUALITIES),
+            default: RunDetectorQualities.GOOD,
         },
     });
 
