@@ -22,7 +22,7 @@
  * @property {number|null} startTime timestamp of the run's start, either trigger start if it exists or o2 start or null
  * @property {number|null} endTime timestamp of the run's end, either trigger end if it exists or o2 end or now (always null if start is null)
  * @property {string|null} environmentId
- * @property {string|null} runQuality
+ * @property {string} runQuality
  * @property {number|null} nDetectors
  * @property {number|null} nFlps
  * @property {number|null} nEpns
