@@ -1225,10 +1225,12 @@ module.exports = {
                         id: 54,
                         run_number: 54,
                         time_o2_start: '2019-08-08 13:00:00',
+                        time_o2_end: '2019-08-08 17:00:00',
                         time_trg_start: '2019-08-08 14:00:00',
+                        time_trg_end: '2019-08-08 16:00:00',
                         environment_id: '1234567890',
                         run_type_id: 12,
-                        run_quality: 'test',
+                        run_quality: 'good',
                         n_detectors: 2,
                         n_flps: 12,
                         n_epns: 12,
@@ -2091,7 +2093,7 @@ module.exports = {
                         time_trg_end: '2019-08-08 13:00:00',
                         environment_id: '1234567890',
                         run_type_id: 12,
-                        run_quality: 'test',
+                        run_quality: 'none',
                         n_detectors: 6,
                         n_flps: 12,
                         n_epns: 12,
