@@ -16,6 +16,7 @@ const PaginationDto = require('./PaginationDto');
 const { TagsFilterDto } = require('./filters/TagsFilterDto.js');
 const { CustomJoi } = require('./CustomJoi.js');
 const { RunDefinition } = require('../../server/services/run/getRunDefinition.js');
+const { RUN_QUALITIES } = require('../enums/RunQualities.js');
 
 const customDateFilterErrorMessages = {
     'date.base': 'Creation date must be a real date and in format YYYY-MM-DD or YYYY/MM/DD',
@@ -57,7 +58,7 @@ const FilterDto = Joi.object({
     triggerValues: CustomJoi.stringArray().items(Joi.string().trim().valid('OFF', 'LTU', 'CTP')),
     runDuration: AmountOperatorComparisonDto,
     environmentIds: Joi.string().trim(),
-    runQualities: CustomJoi.stringArray().items(Joi.string().trim().valid('good', 'bad', 'test')),
+    runQualities: CustomJoi.stringArray().items(Joi.string().trim().valid(...RUN_QUALITIES)),
     nDetectors: AmountOperatorComparisonDto,
     nEpns: AmountOperatorComparisonDto,
     nFlps: AmountOperatorComparisonDto,
