@@ -12,9 +12,10 @@
  * or submit itself to any jurisdiction.
  */
 const Joi = require('joi');
+const { RUN_QUALITIES } = require('../enums/RunQualities.js');
 
 const BodyDto = Joi.object({
-    runQuality: Joi.string().valid('good', 'bad', 'test').optional(),
+    runQuality: Joi.string().valid(...RUN_QUALITIES).optional(),
     lhcBeamEnergy: Joi.number().optional(),
     lhcBeamMode: Joi.string().max(32).optional(),
     lhcBetaStar: Joi.number().optional(),
