@@ -12,12 +12,13 @@
  */
 
 const Joi = require('joi');
-const RunDetectorQuality = require('../enums/DetectorQualities');
+const { RUN_DETECTOR_QUALITIES } = require('../enums/RunDetectorQualities.js');
 const EntityIdDto = require('./EntityIdDto');
+const { RUN_QUALITIES } = require('../enums/RunQualities.js');
 
 const DetectorDto = Joi.object({
     detectorId: EntityIdDto.required(),
-    quality: Joi.string().valid(RunDetectorQuality.Bad, RunDetectorQuality.Good).required(),
+    quality: Joi.string().valid(...RUN_DETECTOR_QUALITIES).required(),
 });
 
 const EorReasonDto = Joi.object({
@@ -33,7 +34,7 @@ const EorReasonDto = Joi.object({
 });
 
 const BodyDto = Joi.object({
-    runQuality: Joi.string().valid('good', 'bad', 'test').optional(),
+    runQuality: Joi.string().valid(...RUN_QUALITIES).optional(),
     eorReasons: Joi.array().min(0).items(EorReasonDto).optional(),
     tags: Joi.array()
         .optional()
