@@ -23,7 +23,7 @@
  * @property {number|null} endTime timestamp of the run's end, either trigger end if it exists or o2 end or now (always null if start is null)
  * @property {string|null} environmentId
  * @property {number|null} runTypeId
- * @property {string|null} runQuality
+ * @property {string} runQuality
  * @property {number|null} nDetectors
  * @property {number|null} nFlps
  * @property {number|null} nEpns
