@@ -12,8 +12,7 @@
  */
 
 import { checkboxFilter } from '../common/filters/checkboxFilter.js';
-
-const RUN_QUALITIES = ['good', 'bad', 'test'];
+import { RUN_QUALITIES } from '../../../domain/enums/RunQualities.js';
 
 /**
  * Returns a panel to be used by user to filter runs by quality
