@@ -20,6 +20,8 @@ import { formatRunType } from '../../utilities/formatting/formatRunType.js';
 import { popover } from '../common/popover/popover.js';
 import { dropdown } from '../common/popover/dropdown.js';
 import { iconCheck, iconX } from '/js/src/icons.js';
+import { RUN_QUALITIES, RunQualities } from '../../domain/enums/RunQualities.js';
+import { RUN_DETECTOR_QUALITIES, RunDetectorQualities } from '../../domain/enums/RunDetectorQualities.js';
 
 /**
  * Method to retrieve the information for a specific run
@@ -326,35 +328,35 @@ const formatEorReasonsCell = (model, eorReasons) => {
  * Formatting for the run quality.
  *
  * @param {Model} model The general model object
- * @param {('good'| 'bad' | 'test' | null)} runQuality The quality of the run
+ * @param {('good'| 'bad' | 'test' | 'none')} runQuality The quality of the run
  * @param {Run} run the run for which run quality must be updated
  * @return {Vnode | String} A box with a color corresponding to the quality or '-' when the runQuality is null
  */
 const formatRunQuality = (model, runQuality, run) => {
     if (model.runs.detailsModel.isEditModeEnabled && (run.timeTrgEnd || run.timeO2End)) {
         return [
-            h('select.form-control.w-15#runQualitySelect', {
-                onchange: ({ target }) => {
-                    model.runs.detailsModel.runChanges = { key: 'runQuality', value: target.value };
+            h(
+                'select.form-control.w-15#runQualitySelect',
+                {
+                    onchange: ({ target }) => {
+                        model.runs.detailsModel.runChanges = { key: 'runQuality', value: target.value };
+                    },
                 },
-            }, [
-                h('option#runQualityGood', {
-                    value: 'good', selected: model.runs.detailsModel.run.payload.runQuality === 'good',
-                }, 'good'),
-                h('option#runQualityBad', {
-                    value: 'bad', selected: model.runs.detailsModel.run.payload.runQuality === 'bad',
-                }, 'bad'),
-                h('option#runQualityTest', {
-                    value: 'test', selected: model.runs.detailsModel.run.payload.runQuality === 'test',
-                }, 'test'),
-            ]),
+                RUN_QUALITIES.map((quality) => h(
+                    `option#run-quality-${quality}`,
+                    {
+                        value: quality, selected: model.runs.detailsModel.run.payload.runQuality === quality,
+                    },
+                    quality,
+                )),
+            ),
         ];
     } else if (runQuality) {
         return h('.badge.white', {
             class: (() => {
-                if (runQuality === 'good') {
+                if (runQuality === RunQualities.GOOD) {
                     return 'bg-success';
-                } else if (runQuality === 'bad') {
+                } else if (runQuality === RunQualities.BAD) {
                     return 'bg-danger';
                 }
                 return 'bg-gray-darker';
@@ -383,7 +385,7 @@ const formatRunDetectors = (run, detailsModel) => {
             const actualQuality = detailsModel.isEditModeEnabled
                 ? detailsModel.getEditionRunDetectorQuality(id) || quality
                 : quality;
-            const qualityIsBad = actualQuality === 'bad';
+            const qualityIsBad = actualQuality === RunDetectorQualities.BAD;
             const color = qualityIsBad ? 'danger' : 'success';
             return h(
                 `span.detector-badge.flex-row.justify-center.lh1.g1.br2.pv1.ph2.b1.b-${color}.${color}`,
@@ -407,7 +409,7 @@ const formatRunDetectors = (run, detailsModel) => {
                             name,
                             h(
                                 '.flex-row.g2.items-center',
-                                ['Good', 'Bad'].map((qualityOption) => h('label.form-check-label.flex-row.g2', [
+                                RUN_DETECTOR_QUALITIES.map((qualityOption) => h('label.form-check-label.flex-row.g2', [
                                     h(
                                         `input#detector-quality-${id}-${qualityOption.toLowerCase()}`,
                                         {
