@@ -14,6 +14,7 @@
 import { detailsList } from '../Detail/detailsList.js';
 import { h } from '/js/src/index.js';
 import { formatTimestamp } from '../../utilities/formatting/formatTimestamp.js';
+import { formatEmail } from '../../utilities/formatting/formatEmail.js';
 
 /**
  * The method to correctly format the current values given from the backend
@@ -49,7 +50,7 @@ const activeFields = (model) => ({
         name: 'Email',
         visible: true,
         size: 'cell-l',
-        format: (email) => model.tags.isEditModeEnabled
+        format: (emails) => model.tags.isEditModeEnabled
             ? h('input.form-control.w-60', {
                 placeholder: 'Enter email groups... ("," separation)',
                 value: model.tags.tagChanges['email'],
@@ -57,7 +58,7 @@ const activeFields = (model) => ({
                     model.tags.setTagChange('email', e.target.value);
                 },
             })
-            : email || '-',
+            : formatTagEmails(emails),
     },
     updatedAt: {
         name: 'Last modified',
@@ -110,5 +111,18 @@ const activeFields = (model) => ({
  */
 export const tagDetail = (model, post) => {
     const postFields = activeFields(model);
-    return detailsList(postFields, post, 'Tag');
+    return detailsList(postFields, post, 'tag');
 };
+
+/**
+ * Format a given emails list (comma separated values)
+ *
+ * @param {string} emails the comma separated list of tags
+ * @return {Component} the emails display
+ */
+const formatTagEmails = (emails) => emails
+    ? h(
+        '.flex-row.justify-end.flex-wrap.gc3.gr1',
+        emails.split(',').map(formatEmail),
+    )
+    : '-';
