@@ -12,6 +12,7 @@
  */
 import { NonOverlappingSegmentsMerger } from '../../utilities/statistics/NonOverlappingSegmentsMerger.js';
 import { MAX_RUN_DURATION } from '../run/constants.js';
+import { RunQualities } from '../../domain/enums/RunQualities.js';
 
 /**
  * @typedef LhcFillStatistics
@@ -90,7 +91,7 @@ export class LhcFillStatisticsExtractor {
         const start = Math.max(startTime, this._stableBeamsStart);
         const end = Math.min(endTime, this._stableBeamsEnd);
 
-        if (runQuality !== 'good' || end - start < 0) {
+        if (runQuality !== RunQualities.GOOD || end - start < 0) {
             return;
         }
 
