@@ -16,7 +16,7 @@ const { ConflictError } = require('../../errors/ConflictError.js');
 const RunRepository = require('../../../database/repositories/RunRepository.js');
 const RunDetectorsRepository = require('../../../database/repositories/RunDetectorsRepository.js');
 const { utilities: { TransactionHelper } } = require('../../../database');
-const RunDetectorQuality = require('../../../domain/enums/DetectorQualities.js');
+const { RunDetectorQualities } = require('../../../domain/enums/RunDetectorQualities.js');
 
 /**
  * Create a run in the database and return the auto generated id
@@ -38,7 +38,7 @@ exports.createRun = async (newRun, detectors, transaction) => {
         await TransactionHelper.provide(async () => RunDetectorsRepository.insertMany(detectors.map((detector) => ({
             runNumber,
             detectorId: detector.id,
-            quality: RunDetectorQuality.Good,
+            quality: RunDetectorQualities.GOOD,
         }))), { transaction });
     }
     const { id } = await TransactionHelper.provide(async () => RunRepository.insert(newRun), { transaction });
