@@ -21,8 +21,8 @@ const ReasonTypeRepository = require('../../../database/repositories/ReasonTypeR
 const { createLog } = require('../log/createLog.js');
 const { BadParameterError } = require('../../errors/BadParameterError.js');
 const { getTagsByText } = require('../tag/getTagsByText.js');
-const { getRunDefinition, RunDefinition } = require('./getRunDefinition.js');
 const { formatServerDate } = require('../../utilities/formatServerDate.js');
+const { RunQualities } = require('../../../domain/enums/RunQualities.js');
 
 /**
  * Method to remove existing reason_type - run_id from `eor_reasons` table and insert new ones
@@ -72,11 +72,11 @@ const logQualityChange = async (runNumber, previousQuality, newQuality, transact
     }
     textParts.push(`on ${formatServerDate()} at ${formatServerDate()}`);
 
-    // Do not tag log for test quality
-    const untaggedLogQuality = 'test';
+    // Do not tag log only for good and bad qualities
+    const TAGGED_LOG_QUALITIES = [RunQualities.GOOD, RunQualities.BAD];
 
     let tags = [];
-    if (previousQuality !== untaggedLogQuality && newQuality !== untaggedLogQuality) {
+    if (TAGGED_LOG_QUALITIES.includes(previousQuality) && TAGGED_LOG_QUALITIES.includes(newQuality)) {
         tags = (await getTagsByText(['DPG', 'RC'])).map(({ text }) => text);
     }
 
@@ -126,13 +126,6 @@ exports.updateRun = async (identifier, runPatch, tags, eorReasons, user, transac
         && runPatch.runQuality !== runModel.runQuality
     ) {
         throw new BadParameterError('Run quality can not be updated on a run that has not ended yet');
-    }
-
-    // When a physics run ends, set its quality to good as default to avoid deletion by mistake
-    if (!runEnd && patchEnd && !runPatch.runQuality) {
-        if (getRunDefinition(runModel) === RunDefinition.Physics) {
-            runPatch.runQuality = 'good';
-        }
     }
 
     // Store the run quality to create a log if it changed
