@@ -17,7 +17,7 @@ const { repositories: { TagRepository }, utilities: { QueryBuilder } } = require
  * Return all the tags that belongs to the given list of text (not existing texts are ignored)
  *
  * @param {string[]} texts the list of texts for which tags must be retrieved
- * @param {function|null} qbConfiguration function called with the tags find query builder as parameter to add specific configuration to the
+ * @param {function|null} [qbConfiguration] function called with the tags find query builder as parameter to add specific configuration to the
  *     query
  * @returns {Promise<SequelizeTag[]>} Promise resolving with the list of tags
  */
